pcap: Make yappcap optional
Move starpy into requirements.txt and make extras.txt optional.

This makes yappcap optional.  As individual tests check for it anyway, it's
failure should not force the test suite venv install to fail.

The pjsip rls tests had a python lib dependency on yappcap that wasn't really
required.  Resolved by splitting pcap.py into pcap_proxy.py where the latter
has no yappcap dependency.

Fixes: #70
