+++ extras.txt
@@ -1,2 +1,1 @@
-https://github.com/asterisk/starpy/archive/refs/heads/1.1.zip
 https://github.com/asterisk/yappcap/archive/refs/tags/v0.0.3-py3.zip
