+++ requirements.txt
@@ -26,3 +26,4 @@ txaio==22.2.1
 typing_extensions==4.1.1
 urllib3==1.26.9
 zope.interface==5.4.0
+https://github.com/asterisk/starpy/archive/refs/heads/1.1.zip
