+++ tests/channels/pjsip/subscriptions/rls/lists_of_lists/nominal/mwi/batched/test-config.yaml
@@ -14,7 +14,6 @@ properties:
         - buildoption: 'TEST_FRAMEWORK'
         - python: 'twisted'
         - python: 'starpy'
-        - python: 'yappcap'
         - python: 'lxml'
         - asterisk: 'res_pjsip'
         - asterisk: 'res_pjsip_pubsub'
